+++ android/src/main/kotlin/com/fluttercandies/photo_manager/constant/Methods.kt
@@ -89,6 +89,7 @@ class Methods {
         const val moveAssetsToPath = "moveAssetsToPath"
         const val removeNoExistsAssets = "removeNoExistsAssets"
         const val getColumnNames = "getColumnNames"
+        const val updateCreationDate = "updateCreationDate"
 
         private val needMediaLocationMethods = arrayOf(
             getLatLng,