+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/PhotoManager.kt
@@ -290,6 +290,16 @@ class PhotoManager(private val context: Context) {
         }
     }
 
+    fun updateCreationDate(assetId: String, timestamp: Long, resultHandler: ResultHandler) {
+        try {
+            val result = dbUtils.updateCreationDate(context, assetId, timestamp)
+            resultHandler.reply(result)
+        } catch (e: Exception) {
+            LogUtils.error(e)
+            resultHandler.reply(false)
+        }
+    }
+
     fun removeAllExistsAssets(resultHandler: ResultHandler) {
         val result = dbUtils.removeAllExistsAssets(context)
         resultHandler.reply(result)